// config/site.config.d.ts
declare const siteConfig: {
  baseDirectory: string
  datetimeFormat: string
  email: string
  footer: string
  googleFontLinks: string[]
  googleFontMono: string
  googleFontSans: string
  icon: string
  kvPrefix: string
  links: Array<{ link: string; name: string }>
  maxItems: number
  protectedRoutes: string[]
  title: string
  userPrincipalName: string
}

export default siteConfig